backend/requirements.txt
# Flask Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
PyJWT==2.8.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
twilio>=8.0.0
eventlet==0.34.2
bcrypt>=4.0.0
groq>=0.4.0

